add vendor and move main.go out
